Nineth, Modal popup window
Continue working on "department.js" file, adding a Modal to show when Add and Update a Department
The Modal only show one field, and it needs update in status or memory using the "onchangeDepartmentName"
To Add will use a new button at top with text "Add Department", to update the same icon with "pencil" at right of each record. As well the delete at end.
The addClick and editClick are called , and those are showed in a new "Modal" and depend if data is empty or with data, will recall "CreateClick" and "UpdateClick" to use the API
At the end the "Delete" doesn't need the "Modal" but ask to be sure the process and call "deleteClick" to use the API

#20211010-step15