Tenth, Employee Details Screen
Base on "department.js" file, we make some addings to "employee.js"
Process similar to "department.js" but the new ones are: Department is a in the Modal is a Drop-down, based on data of Department table and
the "DateOfJoining" field is "Date" type, but there is an error pending for solve, it is when edit the Employee, this date is not backing up the original.

The Previous error in "Edit", if not select a correct date, breaks the API connection.

#20211011-step16